Move actions to schoolings controller
