--- app/views/students/confirm_abrogation.html.haml
@@ -1,12 +1,0 @@
-.fr-col-md-6
-  = dsfr_alert(type: :warning, title: "Êtes-vous certain(e) de vouloir abroger cette décision d'attribution ?")
-
-.fr-col-md-6.fr-mt-3w
-  %ul
-    %li
-      Etudiant :
-      %span= @student.full_name
-
-  .fr-btns-group.fr-btns-group--inline.fr-mt-3w
-    = link_to "Annuler", class_student_path(@classe, @student), class: 'fr-btn fr-btn--secondary'
-    = button_to "Confirmer l'abrogation", abrogate_decision_class_student_path(@classe, @student), method: :delete, class: 'fr-btn fr-btn--danger fr-btn--secondary'
