+++ config/locales/fr.yml
@@ -38,6 +38,8 @@ fr:
       not_found: "La classe que vous avez demandée n'existe pas."
     students:
       not_found: "L'élève demandé n'est pas ou plus dans cette classe."
+    schoolings:
+      not_found: "La scolarité demandée n'est plus disponible ou a été supprimée."
     pfmps:
       not_found: "La PFMP demandée n'est plus disponible ou a été supprimée."
   activerecord:
