+++ config/routes.rb
@@ -49,6 +49,10 @@ Rails.application.routes.draw do
     end
 
     resources :schoolings, only: [] do
+      member do
+        get "confirm_abrogation"
+        delete "abrogate_decision"
+      end
       resources :pfmps, except: :index do
         member do
           post "validate"
@@ -59,10 +63,6 @@ Rails.application.routes.draw do
     end
 
     resources :students, only: %i[show] do
-      member do
-        get "confirm_abrogation"
-        delete "abrogate_decision"
-      end
       resources :ribs do
         member do
           get "confirm_deletion"
