+++ spec/requests/schoolings_controller_spec.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.describe SchoolingsController do
+  let(:schooling) { create(:schooling, :with_attributive_decision) }
+  let(:student) { schooling.student }
+  let(:user) { create(:user, :director, :with_selected_establishment, establishment: student.classe.establishment) }
+
+  before { sign_in(user) }
+
+  describe "DEL /abrogate_decision" do
+    it "enqueues a job to generate the abrogation document" do
+      expect(GenerateAbrogationDecisionJob).to receive(:perform_now).with(schooling) # rubocop:disable RSpec/MessageSpies
+      delete abrogate_decision_class_schooling_path(class_id: schooling.classe.id, id: schooling.id)
+    end
+  end
+end
